Removed redundant code from RSAStandardPrivateKey

+++ src/claire/util/crypto/publickey/rsa/RSAStandardPrivateKey.java
@@ -95,7 +95,7 @@ public class RSAStandardPrivateKey
 	{
 		Bits.intToBytes(len, bytes, offset);
 		mod.export(bytes, offset += 4);
-		exp.export(bytes, (int) (offset + mod.exportSize()));
+		exp.export(bytes, offset + mod.exportSize());
 	}
 
 	public int exportSize()
@@ -122,7 +122,7 @@ public class RSAStandardPrivateKey
 		{
 			int len = Bits.intFromBytes(data, start);
 			UInt mod = UInt.factory.resurrect(data, start += 4);
-			return new RSAStandardPrivateKey(mod, UInt.factory.resurrect(data, (int) (start + mod.exportSize())), len);
+			return new RSAStandardPrivateKey(mod, UInt.factory.resurrect(data, start + mod.exportSize()), len);
 		}
 
 		public RSAStandardPrivateKey resurrect(IIncomingStream stream)throws InstantiationException, IOException
